+++ cmd/indago/main.go
@@ -31,7 +31,7 @@ import (
 )
 
 var (
-	version = "1.3.0"
+	version = "1.3.1"
 	cfgFile string
 	config  *types.Config
 )